@@ -1,62 +1,109 @@
 <svg
 	id="title"
 	class="svg-class"
-	viewBox="0 0 1238 839"
+	viewBox="0 0 950 700"
 	fill="none"
 	xmlns="http://www.w3.org/2000/svg"
 >
 	<path
-		d="M200.538 534.168L5.77344 640.704L5.94954 589.104L164.637 505.245L165.392 518.448L6.47785 434.305L6.65259 383.106L200.682 492.168L296.281 492.495L296.138 534.494L200.538 534.168Z"
+		d="M6.44629 512.081L6.54955 481.825L108.602 421.662L108.611 419.105L6.97131 358.246L7.07457 327.991L135.072 407.263L224.986 407.57L224.896 433.991L134.982 433.684L6.44629 512.081Z"
 	/>
 	<path
-		d="M5.13965 826.309L5.28162 784.709L254.88 785.561L255.377 639.962L295.777 640.1L295.138 827.299L5.13965 826.309Z"
+		d="M224.088 670.571L5.90771 669.826L5.99788 643.406L200.741 644.07L201.087 542.65L224.525 542.73L224.088 670.571Z"
 	/>
 	<path
-		d="M200.682 492.168L296.281 492.494L296.138 534.494L200.539 534.168L5.77398 640.704L5.95009 589.104L152.535 511.64L6.4784 434.305L6.65314 383.105L200.682 492.168ZM141.842 511.636L0.960362 586.086L0.745117 649.154L201.809 539.172L301.121 539.511L301.299 487.512L201.999 487.173L1.68222 374.576L1.46812 437.31L141.842 511.636Z"
+		fill-rule="evenodd"
+		clip-rule="evenodd"
+		d="M1.41602 521.006L1.5595 478.963L100.97 420.357L1.96184 361.075L2.10533 319.032L136.503 402.268L230.003 402.587L229.879 439.008L136.378 438.689L1.41602 521.006ZM134.982 433.684L224.896 433.991L224.986 407.57L135.072 407.263L7.07478 327.991L6.97152 358.246L108.611 419.105L108.602 421.662L6.54976 481.825L6.4465 512.081L134.982 433.684Z"
+		fill="transparent"
+		class="fill-border"
 	/>
 	<path
-		d="M0.123047 831.292L0.299151 779.692L249.898 780.544L250.395 634.945L300.794 635.117L300.121 832.316L0.123047 831.292ZM254.881 785.561L5.28206 784.709L5.14008 826.309L295.138 827.299L295.777 640.1L255.378 639.962L254.881 785.561Z"
+		fill-rule="evenodd"
+		clip-rule="evenodd"
+		d="M229.072 675.588L0.891113 674.809L1.01541 638.389L195.759 639.053L196.105 537.633L229.542 537.747L229.072 675.588ZM200.742 644.07L5.99832 643.406L5.90815 669.826L224.089 670.571L224.525 542.73L201.088 542.65L200.742 644.07Z"
+		fill="transparent"
+		class="fill-border"
 	/>
 	<path
-		d="M1232.87 7.38816L1008.91 230.225L1012.95 217.839L1222.15 217.753L1222.01 259.753L922.812 258.731L922.818 257.131L1146.77 37.0945L1144.34 46.6862L932.736 46.7641L932.878 5.16429L1232.88 6.18815L1232.87 7.38816Z"
+		d="M635.2 5.22268L853.38 5.9673L853.293 31.5353L681.582 149.842L681.574 151.973L852.88 152.557L852.79 178.978L634.609 178.233L634.697 152.665L806.836 33.9336L806.843 31.8029L635.111 31.2168L635.2 5.22268Z"
 	/>
 	<path
-		d="M1221.02 550.806L924.351 428.085L924.353 427.446L1221.85 306.752L1221.72 345.405L1136.31 378.975L1135.97 479.599L1221.14 514.07L1221.02 550.806ZM1105.66 467.357L1105.92 391.01L1048.98 413.496C1041.73 416.24 1034.59 418.878 1027.55 421.409C1020.52 423.941 1012.1 426.468 1002.29 428.99C1012.08 431.579 1020.48 434.056 1027.5 436.423C1034.31 438.789 1041.44 441.475 1048.88 444.482L1105.66 467.357Z"
+		d="M797.207 283.59L796.965 354.781L852.13 372.108L852.038 399.207L638.993 329.632L639.07 306.927L852.581 239.98L852.489 266.933L797.207 283.59ZM773.406 347.376L773.599 290.687L680.926 318.496L773.406 347.376Z"
 	/>
 	<path
-		d="M1237.86 9.47369L1033.48 212.83L1227.17 212.751L1226.99 264.77L917.795 263.714L917.825 255.027L1124.78 51.6935L927.718 51.7659L927.894 0.147257L1237.89 1.20525L1237.86 9.47369ZM1137 46.6889L922.817 257.131L922.812 258.731L1222.01 259.753L1222.15 217.753L1021.36 217.835L1232.87 7.38816L1232.88 6.18815L932.877 5.16429L932.735 46.7641L1137 46.6889Z"
+		fill-rule="evenodd"
+		clip-rule="evenodd"
+		d="M630.217 0.205644L858.398 0.984399L858.285 34.1684L694.495 147.017L857.898 147.574L857.773 183.995L629.593 183.216L629.706 150.034L793.934 36.7589L630.095 36.1998L630.217 0.205644ZM806.844 31.803L806.836 33.9336L634.697 152.665L634.61 178.233L852.79 178.978L852.881 152.557L681.575 151.973L681.582 149.842L853.294 31.5353L853.381 5.96731L635.2 5.22268L635.112 31.2169L806.844 31.803Z"
+		fill="transparent"
+		class="fill-border"
 	/>
 	<path
-		d="M1225.99 558.275L919.34 431.423L919.365 424.074L1226.88 299.318L1226.71 348.816L1141.3 382.386L1140.98 476.233L1226.15 510.705L1225.99 558.275ZM1135.97 479.599L1136.31 378.975L1221.72 345.405L1221.85 306.752L924.353 427.446L924.351 428.085L1221.02 550.806L1221.14 514.07L1135.97 479.599ZM1050.79 418.16L1050.75 418.173C1043.48 420.926 1036.31 423.573 1029.25 426.114C1026.57 427.076 1023.71 428.034 1020.67 428.989C1023.66 429.901 1026.47 430.799 1029.1 431.685L1029.12 431.693L1029.15 431.7C1036.04 434.096 1043.24 436.812 1050.75 439.846L1100.69 459.962L1100.9 398.37L1050.79 418.16ZM1105.92 391.01L1105.66 467.357L1048.88 444.482C1041.44 441.475 1034.31 438.789 1027.5 436.423C1022.88 434.865 1017.66 433.259 1011.84 431.604C1008.82 430.746 1005.64 429.874 1002.29 428.99C1005.64 428.13 1008.82 427.269 1011.84 426.408C1017.68 424.744 1022.92 423.078 1027.55 421.409C1034.59 418.878 1041.73 416.24 1048.98 413.496L1105.92 391.01Z"
+		fill-rule="evenodd"
+		clip-rule="evenodd"
+		d="M852.49 266.933L797.208 283.59L796.965 354.781L852.131 372.108L852.039 399.207L638.994 329.632L639.071 306.927L852.582 239.98L852.49 266.933ZM801.978 351.114L857.144 368.441L857.015 406.092L633.981 333.255L634.084 303.251L857.605 233.165L857.478 270.652L802.196 287.309L801.978 351.114ZM773.6 290.687L680.927 318.496L773.407 347.376L773.6 290.687ZM768.43 340.584L768.577 297.415L698.005 318.591L768.43 340.584Z"
+		fill="transparent"
+		class="fill-border"
 	/>
-	<path d="M739.544 237H781.144V527H739.544V237Z" />
+	<path d="M533.254 186.719V400H504.982V186.719H533.254Z" />
 	<path
-		d="M437.2 527V237H517.6C546.4 237 570.8 241.4 590.8 250.2C610.8 258.733 627.067 270.333 639.6 285C652.134 299.4 661.2 315.4 666.8 333C672.667 350.6 675.6 368.2 675.6 385.8C675.6 407.4 671.734 427 664 444.6C656.534 461.933 646 476.733 632.4 489C619.067 501.267 603.734 510.733 586.4 517.4C569.067 523.8 550.8 527 531.6 527H437.2ZM478.8 486.6H523.6C544.667 486.6 563.334 482.6 579.6 474.6C595.867 466.6 608.667 454.867 618 439.4C627.6 423.933 632.4 405.133 632.4 383C632.4 364.867 629.334 349.267 623.2 336.2C617.334 322.867 609.467 311.933 599.6 303.4C589.734 294.6 578.667 288.067 566.4 283.8C554.134 279.533 541.6 277.4 528.8 277.4H478.8V486.6Z"
-	/>
-	<path d="M734.544 232H786.144V532H734.544V232ZM739.544 237V527H781.144V237H739.544Z" />
-	<path
-		d="M432.2 532V232H517.6C546.871 232 571.991 236.467 592.788 245.612C613.387 254.405 630.3 266.426 643.387 281.735C656.337 296.619 665.742 313.199 671.554 331.451C677.575 349.522 680.6 367.645 680.6 385.8C680.6 408 676.625 428.291 668.585 446.594C660.847 464.553 649.905 479.94 635.767 492.697C621.963 505.392 606.093 515.183 588.195 522.067L588.164 522.079L588.132 522.091C570.223 528.703 551.368 532 531.6 532H432.2ZM632.4 489C646 476.733 656.534 461.933 664 444.6C671.734 427 675.6 407.4 675.6 385.8C675.6 368.2 672.667 350.6 666.8 333C661.2 315.4 652.134 299.4 639.6 285C627.067 270.333 610.8 258.733 590.8 250.2C570.8 241.4 546.4 237 517.6 237H437.2V527H531.6C550.8 527 569.067 523.8 586.4 517.4C603.734 510.733 619.067 501.267 632.4 489ZM613.735 436.79L613.752 436.763C622.764 422.244 627.4 404.403 627.4 383C627.4 365.418 624.426 350.579 618.674 338.325L618.648 338.269L618.624 338.214C613.021 325.479 605.573 315.176 596.329 307.182L596.301 307.157L596.272 307.131C586.888 298.761 576.391 292.569 564.758 288.522C553.011 284.437 541.033 282.4 528.8 282.4H483.8V481.6H523.6C544.026 481.6 561.917 477.725 577.394 470.113C592.804 462.534 604.883 451.46 613.719 436.817L613.735 436.79ZM478.8 486.6V277.4H528.8C541.6 277.4 554.134 279.533 566.4 283.8C578.667 288.067 589.734 294.6 599.6 303.4C609.467 311.933 617.334 322.867 623.2 336.2C629.334 349.267 632.4 364.867 632.4 383C632.4 405.133 627.6 423.933 618 439.4C608.667 454.867 595.867 466.6 579.6 474.6C563.334 482.6 544.667 486.6 523.6 486.6H478.8Z"
-	/>
-	<path
-		d="M1131.66 833.5C1118.13 833.5 1106.46 830.816 1096.66 825.45C1086.86 819.85 1078.93 812.733 1072.86 804.1L1076.36 793.95V830H1041.01V558.05H1076.36V715.9L1075.31 702.95C1081.15 695.483 1089.31 689.416 1099.81 684.75C1110.55 680.083 1122.68 677.75 1136.21 677.75C1149.28 677.75 1161.18 680.9 1171.91 687.2C1182.65 693.266 1191.28 702.016 1197.81 713.45C1204.35 724.883 1207.61 738.533 1207.61 754.4C1207.61 770.966 1204 785.2 1196.76 797.1C1189.76 808.766 1180.55 817.75 1169.11 824.05C1157.68 830.35 1145.2 833.5 1131.66 833.5ZM1123.96 803.75C1133.06 803.75 1141.23 801.65 1148.46 797.45C1155.7 793.016 1161.3 787.066 1165.26 779.6C1169.46 772.133 1171.56 763.733 1171.56 754.4C1171.56 745.066 1169.46 736.783 1165.26 729.55C1161.06 722.316 1155.35 716.6 1148.11 712.4C1141.11 707.966 1133.06 705.75 1123.96 705.75C1112.3 705.75 1102.15 708.783 1093.51 714.85C1084.88 720.916 1079.16 728.966 1076.36 739V769.8C1079.4 779.6 1085 787.766 1093.16 794.3C1101.56 800.6 1111.83 803.75 1123.96 803.75Z"
+		d="M284.705 400V186.719H334.07C342.371 186.816 350.184 187.695 357.508 189.355C364.93 190.918 371.766 193.164 378.016 196.094C386.707 200.098 394.373 205.371 401.014 211.914C407.752 218.359 413.123 225.83 417.127 234.326C420.74 241.553 423.475 249.463 425.33 258.057C427.283 266.65 428.309 275.83 428.406 285.596V301.27C428.406 310.645 427.479 319.531 425.623 327.93C423.865 336.328 421.277 344.043 417.859 351.074C414.344 358.398 409.9 365.039 404.529 370.996C399.158 376.953 393.152 382.031 386.512 386.23C379.48 390.625 371.521 393.994 362.635 396.338C353.846 398.682 344.324 399.902 334.07 400H284.705ZM312.244 208.984V377.881H334.07C341.883 377.783 348.963 376.807 355.311 374.951C361.756 373.096 367.469 370.459 372.449 367.041C377.723 363.525 382.264 359.229 386.072 354.15C389.979 348.975 393.055 343.213 395.301 336.865C397.254 331.689 398.719 326.123 399.695 320.166C400.672 314.111 401.209 307.812 401.307 301.27V285.303C401.209 278.662 400.623 272.314 399.549 266.26C398.572 260.107 397.059 254.395 395.008 249.121C392.371 242.285 388.807 236.133 384.314 230.664C379.92 225.098 374.5 220.605 368.055 217.188C363.465 214.648 358.338 212.695 352.674 211.328C347.01 209.863 340.809 209.082 334.07 208.984H312.244Z"
 	/>
 	<path
-		d="M895.477 833.5C881.71 833.5 869.227 830.583 858.027 824.75C846.827 818.917 837.96 810.4 831.427 799.2C824.894 787.767 821.627 773.65 821.627 756.85C821.627 740.517 825.01 726.517 831.777 714.85C838.777 702.95 847.877 693.85 859.077 687.55C870.51 681.017 882.877 677.75 896.177 677.75C910.177 677.75 921.727 680.667 930.827 686.5C939.927 692.333 946.927 699.333 951.827 707.5L949.727 713.1L953.227 683.7H985.777V830H950.427V792.9L953.927 801.65C952.994 803.517 951.01 806.2 947.977 809.7C944.944 813.2 940.977 816.817 936.077 820.55C931.177 824.05 925.227 827.083 918.227 829.65C911.46 832.217 903.877 833.5 895.477 833.5ZM904.927 804.45C916.594 804.45 926.394 801.417 934.327 795.35C942.494 789.05 947.86 780.533 950.427 769.8V739.7C947.627 730.133 941.91 722.2 933.277 715.9C924.877 709.6 915.077 706.45 903.877 706.45C895.477 706.45 887.777 708.55 880.777 712.75C873.777 716.717 868.177 722.433 863.977 729.9C859.777 737.133 857.677 745.533 857.677 755.1C857.677 764.433 859.777 772.833 863.977 780.3C868.41 787.767 874.244 793.717 881.477 798.15C888.71 802.35 896.527 804.45 904.927 804.45Z"
+		fill-rule="evenodd"
+		clip-rule="evenodd"
+		d="M538.254 181.719V405H499.982V181.719H538.254ZM533.254 186.719H504.982V400H533.254V186.719Z"
+		fill="transparent"
+		class="fill-border"
 	/>
-	<path d="M743.65 558.05H779V830H743.65V558.05Z" />
 	<path
-		d="M1094.22 829.814L1094.18 829.791C1089.53 827.131 1085.25 824.136 1081.36 820.808V835H1036.01V553.05H1081.36V689.909C1086.13 686.159 1091.61 682.923 1097.78 680.181L1097.8 680.173L1097.82 680.164C1109.27 675.185 1122.1 672.75 1136.21 672.75C1150.11 672.75 1162.88 676.108 1174.41 682.867C1185.96 689.407 1195.21 698.822 1202.15 710.969C1209.2 723.299 1212.61 737.841 1212.61 754.4C1212.61 771.703 1208.83 786.87 1201.04 799.688C1193.61 812.071 1183.76 821.687 1171.53 828.429C1159.33 835.152 1146.01 838.5 1131.66 838.5C1117.47 838.5 1104.94 835.682 1094.26 829.835L1094.22 829.814ZM1081.36 696.453C1086.4 691.881 1092.55 687.98 1099.81 684.75C1110.55 680.083 1122.68 677.75 1136.21 677.75C1149.28 677.75 1161.18 680.9 1171.91 687.2C1182.65 693.266 1191.28 702.016 1197.81 713.45C1204.35 724.883 1207.61 738.533 1207.61 754.4C1207.61 770.966 1204 785.2 1196.76 797.1C1189.76 808.766 1180.55 817.75 1169.11 824.05C1157.68 830.35 1145.2 833.5 1131.66 833.5C1118.13 833.5 1106.46 830.816 1096.66 825.45C1090.92 822.17 1085.82 818.369 1081.36 814.048C1079.6 812.335 1077.93 810.541 1076.36 808.665V830H1041.01V558.05H1076.36V701.651C1077.89 699.831 1079.55 698.099 1081.36 696.453ZM1160.88 777.201L1160.91 777.148C1164.65 770.496 1166.56 762.953 1166.56 754.4C1166.56 745.844 1164.65 738.446 1160.94 732.06C1157.18 725.579 1152.08 720.487 1145.6 716.724L1145.52 716.675L1145.44 716.624C1139.32 712.747 1132.21 710.75 1123.96 710.75C1113.21 710.75 1104.09 713.528 1096.39 718.941C1088.87 724.222 1083.91 731.098 1081.36 739.703V769.025C1084.13 777.512 1089.05 784.592 1096.23 790.346C1103.62 795.869 1112.77 798.75 1123.96 798.75C1132.24 798.75 1139.51 796.855 1145.9 793.156C1152.35 789.191 1157.31 783.912 1160.85 777.254L1160.88 777.201ZM1148.46 797.45C1141.23 801.65 1133.06 803.75 1123.96 803.75C1111.83 803.75 1101.56 800.6 1093.16 794.3C1085 787.766 1079.4 779.6 1076.36 769.8V739C1079.16 728.966 1084.88 720.916 1093.51 714.85C1102.15 708.783 1112.3 705.75 1123.96 705.75C1133.06 705.75 1141.11 707.966 1148.11 712.4C1155.35 716.6 1161.06 722.316 1165.26 729.55C1169.46 736.783 1171.56 745.066 1171.56 754.4C1171.56 763.733 1169.46 772.133 1165.26 779.6C1161.3 787.066 1155.7 793.016 1148.46 797.45Z"
+		fill-rule="evenodd"
+		clip-rule="evenodd"
+		d="M279.705 405V181.719H334.1L334.129 181.719C342.732 181.82 350.885 182.731 358.576 184.471C366.318 186.103 373.507 188.46 380.123 191.559C389.308 195.793 397.443 201.382 404.496 208.326C411.655 215.18 417.369 223.128 421.625 232.142C425.424 239.753 428.28 248.037 430.212 256.974C432.248 265.943 433.305 275.471 433.406 285.546L433.406 285.571V301.27C433.406 310.965 432.448 320.206 430.511 328.981C428.678 337.728 425.969 345.825 422.361 353.25C418.643 360.994 413.935 368.031 408.243 374.344C402.573 380.633 396.218 386.008 389.184 390.456L389.173 390.463L389.162 390.47C381.663 395.157 373.234 398.713 363.917 401.171C354.677 403.634 344.736 404.899 334.118 405L334.094 405H279.705ZM362.635 396.338C371.521 393.994 379.48 390.625 386.512 386.23C393.152 382.031 399.158 376.953 404.529 370.996C409.9 365.039 414.344 358.398 417.859 351.074C421.277 344.043 423.865 336.328 425.623 327.93C427.479 319.531 428.406 310.645 428.406 301.27V285.596C428.309 275.83 427.283 266.65 425.33 258.057C423.475 249.463 420.74 241.553 417.127 234.326C413.123 225.83 407.752 218.359 401.014 211.914C394.373 205.371 386.707 200.098 378.016 196.094C371.766 193.164 364.93 190.918 357.508 189.355C350.184 187.695 342.371 186.816 334.07 186.719H284.705V400H334.07C344.324 399.902 353.846 398.682 362.635 396.338ZM353.918 370.149L353.927 370.146C359.914 368.423 365.128 366.001 369.62 362.918L369.648 362.899L369.676 362.881C374.477 359.68 378.603 355.776 382.072 351.15L382.077 351.144L382.081 351.138C385.668 346.387 388.505 341.081 390.587 335.197L390.604 335.148L390.623 335.1C392.449 330.261 393.833 325.019 394.761 319.361C395.693 313.574 396.211 307.532 396.307 301.231V285.34C396.211 278.96 395.648 272.893 394.626 267.133L394.618 267.088L394.611 267.044C393.684 261.208 392.257 255.843 390.348 250.933L390.343 250.92C387.891 244.565 384.592 238.88 380.451 233.838L380.42 233.8L380.39 233.762C376.438 228.756 371.562 224.707 365.712 221.605L365.673 221.584L365.634 221.563C361.477 219.263 356.777 217.462 351.501 216.189L351.461 216.179L351.422 216.169C346.206 214.82 340.418 214.079 334.033 213.984H317.244V372.881H334.039C341.465 372.786 348.074 371.857 353.908 370.152L353.918 370.149ZM334.07 377.881H312.244V208.984H334.07C340.809 209.082 347.01 209.863 352.674 211.328C358.338 212.695 363.465 214.648 368.055 217.188C374.5 220.605 379.92 225.098 384.314 230.664C388.807 236.133 392.371 242.285 395.008 249.121C397.059 254.395 398.572 260.107 399.549 266.26C400.623 272.314 401.209 278.662 401.307 285.303V301.27C401.209 307.812 400.672 314.111 399.695 320.166C398.719 326.123 397.254 331.689 395.301 336.865C393.055 343.213 389.979 348.975 386.072 354.15C382.264 359.229 377.723 363.525 372.449 367.041C367.469 370.459 361.756 373.096 355.311 374.951C348.963 376.807 341.883 377.783 334.07 377.881Z"
+		fill="transparent"
+		class="fill-border"
 	/>
 	<path
-		d="M858.027 824.75C869.227 830.583 881.71 833.5 895.477 833.5C903.877 833.5 911.46 832.217 918.227 829.65C925.227 827.083 931.177 824.05 936.077 820.55C939.705 817.785 942.822 815.085 945.427 812.448C946.34 811.524 947.19 810.608 947.977 809.7C948.888 808.648 949.705 807.67 950.427 806.766V830H985.777V683.7H953.227L950.622 705.58C949.307 703.573 947.858 701.64 946.277 699.781C942.068 694.832 936.918 690.404 930.827 686.5C921.727 680.667 910.177 677.75 896.177 677.75C882.877 677.75 870.51 681.017 859.077 687.55C847.877 693.85 838.777 702.95 831.777 714.85C825.01 726.517 821.627 740.517 821.627 756.85C821.627 773.65 824.894 787.767 831.427 799.2C837.96 810.4 846.827 818.917 858.027 824.75ZM856.611 683.2C868.804 676.236 882.022 672.75 896.177 672.75C910.855 672.75 923.414 675.809 933.525 682.291C938.623 685.558 943.142 689.199 947.059 693.219L948.787 678.7H990.777V835H945.427V819.336C943.495 821.061 941.386 822.791 939.107 824.527L939.046 824.574L938.983 824.619C933.649 828.428 927.288 831.651 919.973 834.335C912.564 837.141 904.376 838.5 895.477 838.5C880.98 838.5 867.693 835.422 855.717 829.185C843.657 822.903 834.104 813.712 827.108 801.719L827.097 801.7L827.086 801.681C820.017 789.311 816.627 774.285 816.627 756.85C816.627 739.821 820.158 724.916 827.452 712.341L827.459 712.328L827.467 712.315C834.877 699.718 844.591 689.965 856.611 683.2ZM934.327 795.35C926.394 801.417 916.594 804.45 904.927 804.45C896.527 804.45 888.71 802.35 881.477 798.15C874.244 793.717 868.41 787.767 863.977 780.3C859.777 772.833 857.677 764.433 857.677 755.1C857.677 745.533 859.777 737.133 863.977 729.9C868.177 722.433 873.777 716.717 880.777 712.75C887.777 708.55 895.477 706.45 903.877 706.45C915.077 706.45 924.877 709.6 933.277 715.9C941.91 722.2 947.627 730.133 950.427 739.7V769.8C947.86 780.533 942.494 789.05 934.327 795.35ZM931.281 791.385C924.362 796.672 915.683 799.45 904.927 799.45C897.425 799.45 890.491 797.591 884.039 793.856C877.555 789.871 872.32 784.541 868.306 777.798C864.583 771.158 862.677 763.632 862.677 755.1C862.677 746.286 864.604 738.777 868.301 732.411L868.318 732.381L868.335 732.351C872.119 725.625 877.086 720.589 883.242 717.1L883.296 717.069L883.349 717.037C889.549 713.318 896.356 711.45 903.877 711.45C914.072 711.45 922.802 714.294 930.277 719.9L930.303 719.92L930.33 719.939C937.934 725.488 942.893 732.303 945.427 740.438V769.197C943.059 778.628 938.334 785.941 931.281 791.385Z"
+		d="M730.434 667V467H753.48V540.828H755.434C757.126 538.224 759.47 534.904 762.465 530.867C765.525 526.766 769.887 523.12 775.551 519.93C781.28 516.674 789.027 515.047 798.793 515.047C811.423 515.047 822.556 518.204 832.191 524.52C841.827 530.835 849.346 539.786 854.75 551.375C860.154 562.964 862.855 576.635 862.855 592.391C862.855 608.276 860.154 622.046 854.75 633.699C849.346 645.288 841.859 654.272 832.289 660.652C822.719 666.967 811.684 670.125 799.184 670.125C789.548 670.125 781.833 668.53 776.039 665.34C770.245 662.085 765.785 658.406 762.66 654.305C759.535 650.138 757.126 646.688 755.434 643.953H752.699V667H730.434ZM753.09 592C753.09 603.328 754.75 613.322 758.07 621.98C761.391 630.574 766.241 637.312 772.621 642.195C779.001 647.013 786.814 649.422 796.059 649.422C805.694 649.422 813.734 646.883 820.18 641.805C826.69 636.661 831.573 629.76 834.828 621.102C838.148 612.378 839.809 602.677 839.809 592C839.809 581.453 838.181 571.948 834.926 563.484C831.736 554.956 826.885 548.217 820.375 543.27C813.93 538.257 805.824 535.75 796.059 535.75C786.684 535.75 778.806 538.126 772.426 542.879C766.046 547.566 761.228 554.142 757.973 562.605C754.717 571.004 753.09 580.802 753.09 592Z"
 	/>
-	<path d="M738.65 553.05H784V835H738.65V553.05ZM743.65 558.05V830H779V558.05H743.65Z" />
+	<path
+		d="M616.493 670.516C606.988 670.516 598.362 668.725 590.614 665.145C582.867 661.499 576.715 656.258 572.157 649.422C567.6 642.521 565.321 634.188 565.321 624.422C565.321 615.828 567.014 608.862 570.399 603.523C573.785 598.12 578.31 593.888 583.974 590.828C589.638 587.768 595.888 585.49 602.724 583.992C609.625 582.43 616.558 581.193 623.524 580.281C632.639 579.109 640.028 578.23 645.692 577.645C651.422 576.993 655.588 575.919 658.192 574.422C660.862 572.924 662.196 570.32 662.196 566.609V565.828C662.196 556.193 659.56 548.706 654.286 543.367C649.078 538.029 641.168 535.359 630.556 535.359C619.553 535.359 610.927 537.768 604.677 542.586C598.427 547.404 594.032 552.547 591.493 558.016L569.618 550.203C573.524 541.089 578.733 533.992 585.243 528.914C591.819 523.771 598.98 520.19 606.728 518.172C614.54 516.089 622.222 515.047 629.774 515.047C634.592 515.047 640.126 515.633 646.376 516.805C652.691 517.911 658.778 520.223 664.638 523.738C670.562 527.254 675.478 532.56 679.384 539.656C683.29 546.753 685.243 556.258 685.243 568.172V667H662.196V646.688H661.024C659.462 649.943 656.858 653.426 653.212 657.137C649.566 660.848 644.716 664.005 638.661 666.609C632.606 669.214 625.217 670.516 616.493 670.516ZM620.009 649.812C629.123 649.812 636.806 648.022 643.056 644.441C649.371 640.861 654.123 636.238 657.313 630.574C660.569 624.91 662.196 618.953 662.196 612.703V591.609C661.22 592.781 659.071 593.855 655.751 594.832C652.496 595.743 648.72 596.557 644.423 597.273C640.191 597.924 636.057 598.51 632.021 599.031C628.049 599.487 624.826 599.878 622.353 600.203C616.363 600.984 610.764 602.254 605.556 604.012C600.412 605.704 596.246 608.276 593.056 611.727C589.931 615.112 588.368 619.734 588.368 625.594C588.368 633.602 591.33 639.656 597.255 643.758C603.244 647.794 610.829 649.812 620.009 649.812Z"
+	/>
+	<path d="M530.141 467V667H507.094V467H530.141Z" />
+	<path
+		fill-rule="evenodd"
+		clip-rule="evenodd"
+		d="M725.434 672V462H758.48V527.846C762.051 523.072 766.993 519.012 773.089 515.578C779.848 511.74 788.553 510.047 798.793 510.047C812.299 510.047 824.404 513.438 834.932 520.338C845.429 527.217 853.53 536.926 859.282 549.262C865.057 561.647 867.855 576.071 867.855 592.391C867.855 608.835 865.059 623.353 859.286 635.803L859.282 635.812C853.535 648.136 845.479 657.869 835.063 664.813L835.053 664.819L835.043 664.826C824.582 671.729 812.571 675.125 799.184 675.125C789.064 675.125 780.422 673.46 773.628 669.72L773.609 669.71L773.59 669.699C767.366 666.202 762.322 662.111 758.683 657.335L758.672 657.32L758.66 657.305C758.333 656.868 758.012 656.439 757.699 656.016V672H725.434ZM755.434 643.953C756.083 645.003 756.839 646.158 757.699 647.418C759.081 649.442 760.735 651.737 762.66 654.305C765.785 658.406 770.245 662.085 776.039 665.34C781.833 668.53 789.548 670.125 799.184 670.125C811.684 670.125 822.719 666.967 832.289 660.652C841.859 654.272 849.346 645.288 854.75 633.699C860.154 622.046 862.855 608.276 862.855 592.391C862.855 576.635 860.154 562.964 854.75 551.375C849.346 539.786 841.827 530.835 832.191 524.52C822.556 518.204 811.423 515.047 798.793 515.047C789.027 515.047 781.28 516.674 775.551 519.93C769.887 523.12 765.525 526.766 762.465 530.867C760.976 532.874 759.648 534.704 758.48 536.357C757.3 538.028 756.285 539.519 755.434 540.828H753.48V467H730.434V667H752.699V643.953H755.434ZM830.152 619.333L830.155 619.323C833.23 611.243 834.809 602.156 834.809 592C834.809 581.968 833.261 573.085 830.259 565.279L830.251 565.258L830.243 565.236C827.345 557.488 823.024 551.563 817.35 547.25L817.327 547.233L817.305 547.216C811.9 543.012 804.938 540.75 796.059 540.75C787.561 540.75 780.783 542.888 775.413 546.889L775.399 546.899L775.386 546.908C769.884 550.951 765.602 556.696 762.639 564.4L762.635 564.412C759.65 572.113 758.09 581.275 758.09 592C758.09 602.845 759.68 612.21 762.737 620.185C765.772 628.039 770.107 633.971 775.647 638.214C781.025 642.272 787.731 644.422 796.059 644.422C804.774 644.422 811.667 642.145 817.082 637.88C822.798 633.364 827.175 627.249 830.148 619.342L830.152 619.333ZM820.18 641.805C813.734 646.883 805.694 649.422 796.059 649.422C786.814 649.422 779.001 647.013 772.621 642.195C766.241 637.312 761.391 630.574 758.07 621.98C754.75 613.322 753.09 603.328 753.09 592C753.09 580.802 754.717 571.004 757.973 562.605C761.228 554.142 766.046 547.566 772.426 542.879C778.806 538.126 786.684 535.75 796.059 535.75C805.824 535.75 813.93 538.257 820.375 543.27C826.885 548.217 831.736 554.956 834.926 563.484C838.181 571.948 839.809 581.453 839.809 592C839.809 602.677 838.148 612.378 834.828 621.102C831.573 629.76 826.69 636.661 820.18 641.805Z"
+		fill="transparent"
+		class="fill-border"
+	/>
+	<path
+		fill-rule="evenodd"
+		clip-rule="evenodd"
+		d="M590.614 665.145C598.362 668.725 606.988 670.516 616.493 670.516C625.217 670.516 632.606 669.214 638.661 666.609C644.716 664.005 649.566 660.848 653.212 657.137C654.718 655.604 656.046 654.11 657.196 652.655C658.831 650.587 660.107 648.598 661.024 646.688H662.196V667H685.243V568.172C685.243 556.258 683.29 546.753 679.384 539.656C675.478 532.56 670.562 527.254 664.638 523.738C658.778 520.223 652.691 517.911 646.376 516.805C640.126 515.633 634.592 515.047 629.774 515.047C622.222 515.047 614.54 516.089 606.728 518.172C598.98 520.19 591.819 523.771 585.243 528.914C579.858 533.115 575.363 538.696 571.76 545.659C571.007 547.113 570.293 548.628 569.618 550.203L591.493 558.016C594.032 552.547 598.427 547.404 604.677 542.586C610.927 537.768 619.553 535.359 630.556 535.359C641.168 535.359 649.078 538.029 654.286 543.367C659.56 548.706 662.196 556.193 662.196 565.828V566.609C662.196 570.32 660.862 572.924 658.192 574.422C655.588 575.919 651.422 576.993 645.692 577.645C640.028 578.23 632.639 579.109 623.524 580.281C616.558 581.193 609.625 582.43 602.724 583.992C595.888 585.49 589.638 587.768 583.974 590.828C578.31 593.888 573.785 598.12 570.399 603.523C567.014 608.862 565.321 615.828 565.321 624.422C565.321 634.188 567.6 642.521 572.157 649.422C576.715 656.258 582.867 661.499 590.614 665.145ZM566.17 600.857C570.018 594.719 575.184 589.893 581.597 586.429C587.701 583.131 594.39 580.701 601.636 579.112C608.681 577.517 615.761 576.254 622.876 575.324L622.887 575.322C632.007 574.149 639.433 573.266 645.152 572.674C650.697 572.042 654.015 571.056 655.7 570.087L655.723 570.074L655.746 570.061C656.478 569.651 657.196 568.967 657.196 566.609V565.828C657.196 557.118 654.836 551.038 650.729 546.881L650.718 546.87L650.707 546.859C646.778 542.832 640.397 540.359 630.556 540.359C620.231 540.359 612.818 542.624 607.729 546.546C601.926 551.02 598.141 555.57 596.028 560.121L594.107 564.259L562.927 553.123L565.022 548.234C569.194 538.5 574.869 530.665 582.165 524.974C589.251 519.432 597.021 515.536 605.452 513.337C613.651 511.152 621.762 510.047 629.774 510.047C634.973 510.047 640.816 510.677 647.269 511.885C654.212 513.106 660.861 515.643 667.199 519.444C674.012 523.49 679.506 529.51 683.764 537.245C688.224 545.347 690.243 555.783 690.243 568.172V672H657.196V660.212C657.058 660.355 656.919 660.498 656.779 660.641C652.574 664.92 647.135 668.408 640.637 671.203C633.791 674.147 625.688 675.516 616.493 675.516C606.346 675.516 596.993 673.601 588.517 669.683L588.501 669.676L588.485 669.669C579.95 665.652 573.081 659.822 567.997 652.195L567.991 652.186L567.985 652.177C562.796 644.32 560.321 634.987 560.321 624.422C560.321 615.216 562.13 607.234 566.17 600.857ZM643.056 644.441C636.806 648.022 629.123 649.812 620.009 649.812C610.829 649.812 603.244 647.794 597.255 643.758C591.33 639.656 588.368 633.602 588.368 625.594C588.368 619.734 589.931 615.112 593.056 611.727C596.246 608.276 600.412 605.704 605.556 604.012C610.764 602.254 616.363 600.984 622.353 600.203C624.827 599.878 628.049 599.487 632.021 599.031C636.057 598.51 640.191 597.924 644.423 597.273C648.72 596.557 652.496 595.743 655.751 594.832C656.26 594.682 656.742 594.53 657.196 594.376C659.703 593.524 661.37 592.601 662.196 591.609V612.703C662.196 618.953 660.569 624.91 657.313 630.574C654.123 636.238 649.371 640.861 643.056 644.441ZM657.196 599.619V612.703C657.196 618.038 655.818 623.141 652.978 628.083L652.968 628.102L652.957 628.121C650.264 632.903 646.209 636.906 640.589 640.092L640.58 640.097L640.57 640.103C635.252 643.15 628.479 644.812 620.009 644.812C611.571 644.812 605.038 642.965 600.075 639.629C595.683 636.577 593.368 632.161 593.368 625.594C593.368 620.639 594.669 617.352 596.727 615.121C596.728 615.12 596.729 615.119 596.73 615.118C599.261 612.381 602.659 610.229 607.119 608.761L607.137 608.755L607.155 608.749C612.025 607.105 617.302 605.904 622.999 605.161L623.005 605.16C625.44 604.84 628.632 604.453 632.591 603.999L632.625 603.995L632.66 603.99C636.739 603.464 640.913 602.872 645.183 602.215L645.214 602.211L645.245 602.205C649.671 601.468 653.63 600.618 657.099 599.647L657.131 599.638L657.162 599.629C657.173 599.625 657.185 599.622 657.196 599.619Z"
+		fill="transparent"
+		class="fill-border"
+	/>
+	<path
+		fill-rule="evenodd"
+		clip-rule="evenodd"
+		d="M535.141 462V672H502.094V462H535.141ZM530.141 467H507.094V667H530.141V467Z"
+		fill="transparent"
+		class="fill-border"
+	/>
 </svg>
 
 <style>
 	.svg-class {
-		width: 70%;
-		height: 70%;
-		margin: auto;
+		width: 60%;
+		height: 60%;
+		margin-left: auto;
+	}
+	@media (max-width: 768px) {
+		.svg-class {
+			width: 100%;
+			height: 100%;
+			margin-left: 10%;
+		}
 	}
 </style>
